requirements.txt
ansible-runner==2.3.4
attrs==23.1.0
blinker==1.6.3
boto3==1.28.72
botocore==1.31.72
click==8.1.7
docutils==0.20.1
flasgger==0.9.7.1
Flask==3.0.0
Flask-Cors==4.0.0
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
lockfile==0.12.2
MarkupSafe==2.1.3
mistune==3.0.2
packaging==23.2
pexpect==4.8.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.4
s3transfer==0.7.0
six==1.16.0
urllib3
Werkzeug==3.0.0
